+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # [Changelog v1.3.0](https://github.com/digas99/schedule-ua/releases/tag/v1.3.0)
 
-***Classes notices!***
-
 ## Popup
 - The abbreviation for the Current/Next class of the day now shows up on the extension icon
 - Added possibility to highlight mouse target cell on the schedule (can be enabled on settings)
+- Added subject code to paco schedule scraper and to the extension itself
+- Added mechanism to load each subject schedule with all possible classes (with message logging each load for the user)
 
 ---